src/main/java/ua/com/juja/leetcode/A_220_Contains_Duplicate_III/Solution.java
package ua.com.juja.leetcode.A_220_Contains_Duplicate_III;

/**
 * Created by serzh on 3/14/16.
 */
public class Solution {
    public boolean containsNearbyAlmostDuplicate(int[] nums, int k, int t) {
        if (nums.length < 2 || t < 0 || k < 1) {
            return false;
        }

        for (int i = 0; i < nums.length - 1; i++) {
            for (int j = i + 1; j < nums.length; j++) {
                if ((j - i) >= k) break;
                if (Math.abs(nums[i] - nums[j]) <= t) {
                    return true;
                }
            }
        }
        return false;
    }
}